♻️ Refactor: 프로필 관련 로직을 훅으로 분리
